==> 0_Pytesseract_Extractor.ipynb <==
from pytesseract import pytesseract

pytesseract.tesseract_cmd = r'C:\Program Files\Tesseract-OCR\tesseract.exe'

import os
from pytesseract import image_to_string
from pdf2image import convert_from_path

# Define the paths
pdf_folder = './UniqueCourses'  # Folder containing the PDF files
output_folder = './TextFiles_Pytesseract'  # Folder to save the text files

# Iterate through all PDF files in the folder
for filename in os.listdir(pdf_folder):
    if filename.endswith(".pdf"):
        pdf_path = os.path.join(pdf_folder, filename)
        text_filename = os.path.splitext(filename)[0] + ".txt"  # Change extension to .txt
        text_path = os.path.join(output_folder, text_filename)

        pages = convert_from_path(pdf_path)
        full_text = ""
        for page in pages:
            text = image_to_string(page, lang='eng')
            full_text += text + "\n"
        
        # Write the extracted text to a text file
        with open(text_path, 'w', encoding='utf-8') as text_file:
            text_file.write(full_text)

    print(f"Extracted text from {filename} and saved to {text_filename}")
