==> 7_cleanCSV.ipynb <==
# Clean CSV

Remove square brackets and single quotes that important terms/phrases in the dataset are wrapped in.

### Input: Apriori_Expansion.csv,   Apriori_Ngrams.csv
### Output: Cleaned_Apriori_Expansion.csv,   Cleaned_Apriori_Ngrams.csv

import pandas as pd

# Load the Apriori.csv file
apriori_ex_df = pd.read_csv('Apriori_Expansion.csv')
apriori_ngram_df = pd.read_csv("Apriori_Ngrams.csv")

# Clean the 'Base' and 'Add' columns by removing brackets and single quotes
apriori_ex_df['Base'] = apriori_ex_df['Base'].str.strip("[]'\"")
apriori_ex_df['Add'] = apriori_ex_df['Add'].str.strip("[]'\"")
apriori_ngram_df['Base'] = apriori_ngram_df['Base'].str.strip("[]'\"")
apriori_ngram_df['Add'] = apriori_ngram_df['Add'].str.strip("[]'\"")

# Save the cleaned DataFrame back to a CSV file
cleaned_file_path_ex = 'Cleaned_Apriori_Expansion.csv'
cleaned_file_path_ngram = 'Cleaned_Apriori_Ngrams.csv'
apriori_ex_df.to_csv(cleaned_file_path_ex, index=False)
apriori_ngram_df.to_csv(cleaned_file_path_ngram, index=False)
